Removed formatting edits
